Change CellArray to cellArrayUtils.
