+++ packages/core/src/view/GraphDataModel.ts
@@ -34,6 +34,7 @@ import VisibleChange from './undoable_changes/VisibleChange';
 import Geometry from './geometry/Geometry';
 import ObjectCodec from '../serialization/ObjectCodec';
 import CodecRegistry from '../serialization/CodecRegistry';
+import { cloneCells, filterCells } from '../util/cellArrayUtils';
 
 import type { CellStyle, FilterFunction } from '../types';
 
@@ -338,7 +339,7 @@ export class GraphDataModel extends EventSource {
   }
 
   filterCells(cells: Cell[], filter: FilterFunction) {
-    return cells.filterCells(filter);
+    return filterCells(filter)(cells);
   }
 
   getRoot(cell: Cell | null = null) {
@@ -1168,7 +1169,7 @@ export class GraphDataModel extends EventSource {
    */
   cloneCell(cell: Cell | null = null, includeChildren: boolean = true): Cell | null {
     if (cell != null) {
-      return [cell].cloneCells(includeChildren)[0];
+      return cloneCells(includeChildren)([cell])[0];
     }
     return null;
   }
